adicionado versao estavel com timer para execucao
