+++ observability/grafana/simple-news-app-grafana/src/main/java/br/com/jhegnerlabs/httpserver/AppHttpServer.java
@@ -1,21 +1,20 @@
 package br.com.jhegnerlabs.httpserver;
 
 import com.sun.net.httpserver.HttpServer;
-import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.prometheus.PrometheusMeterRegistry;
 
 import java.io.IOException;
 import java.io.OutputStream;
 import java.net.InetSocketAddress;
 
-public final class AppHttpServer {
+public class AppHttpServer {
 
     public static void start(PrometheusMeterRegistry prometheusMeterRegistry) {
 
         try {
 
             HttpServer server = HttpServer.create(new InetSocketAddress(8080), 0);
-            server.createContext("/prometheus", httpExchange -> {
+            server.createContext("/metrics", httpExchange -> {
                 String response = prometheusMeterRegistry.scrape();
                 httpExchange.sendResponseHeaders(200, response.getBytes().length);
                 try (OutputStream os = httpExchange.getResponseBody()) {
