+++ observability/grafana/simple-news-app-grafana/src/main/java/br/com/jhegnerlabs/metrics/Metrica.java
@@ -1,5 +1,7 @@
 package br.com.jhegnerlabs.metrics;
 
+import br.com.jhegnerlabs.domain.Autor;
+import br.com.jhegnerlabs.domain.Canal;
 import br.com.jhegnerlabs.domain.Noticia;
 import io.micrometer.core.instrument.Counter;
 import io.micrometer.core.instrument.Tag;
@@ -15,6 +17,26 @@ public class Metrica {
                         Tag.of("autor", noticia.getAutor().getNome().toLowerCase()),
                         Tag.of("mes", String.valueOf(noticia.getDataPublicacao().getMonthValue())),
                         Tag.of("ano", String.valueOf(noticia.getDataPublicacao().getYear()))
+                ))
+                .register(MetricaConfig.configure());
+        counter.increment();
+    }
+
+    public void registraMetrica(MetricaType metrica, Autor autor) {
+        Counter counter = Counter.builder(metrica.getMetrica())
+                .description(metrica.getDescricao())
+                .tags(List.of(
+                        Tag.of("autor", autor.getNome().toLowerCase())
+                ))
+                .register(MetricaConfig.configure());
+        counter.increment();
+    }
+
+    public void registraMetrica(MetricaType metrica, Canal canal) {
+        Counter counter = Counter.builder(metrica.getMetrica())
+                .description(metrica.getDescricao())
+                .tags(List.of(
+                        Tag.of("canal", canal.name().toLowerCase())
                 ))
                 .register(MetricaConfig.configure());
         counter.increment();
