+++ observability/grafana/simple-news-app-grafana/src/main/java/br/com/jhegnerlabs/metrics/MetricaType.java
@@ -2,7 +2,9 @@ package br.com.jhegnerlabs.metrics;
 
 public enum MetricaType {
 
-    NOTICIAS("noticias", "A quantidade de noticias publicadas");
+    NOTICIAS("noticias", "Total de noticias publicadas por determinado autor com a informacao do ano e mes"),
+    FAKE_NEWS("fakenews", "Total de noticias identificadas como fakenews e o autor da noticia"),
+    CANAL_PUBLICACAO("canal", "Total de publicacoes num determinado canal");
 
 
     private final String metrica;
