+++ observability/prometheus/prometheus.yml
@@ -1,0 +1,9 @@
+global:
+    scrape_interval: 10s # How frequently to scrape targets by default
+  
+scrape_configs:
+    - job_name: 'simple_news_app_micrometer'  # The job name is assigned to scraped metrics by default.
+      metrics_path: '/metrics'    # The HTTP resource path on which to fetch metrics from targets.
+      scrape_interval: 5s                     # How frequently to scrape targets from this job.
+      static_configs:                         # A static_config allows specifying a list of targets and a common label set for them
+        - targets: ['192.168.0.10:8080']
